calmness-website/calmness-website-vitrine/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.0.1
aiofiles==23.2.1
# Authentication & OAuth
authlib==1.2.1
httpx==0.25.2
# 2FA & SMS
pyotp==2.9.0
twilio==8.10.0
qrcode==7.4.2
# Email
fastapi-mail==1.4.1
# Security
cryptography==41.0.7
# Database
sqlalchemy==2.0.23
alembic==1.12.1
